+++ mustachejs/tests/project/project/views.py
@@ -1,0 +1,4 @@
+from django.utils.translation import ugettext as _
+
+def f():
+    return _('View message')